requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2
django-cors-headers==4.2.0
django-filter==23.2
django-guardian==2.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
pytz==2023.3.post1
PyYAML==6.0.1
setuptools==70.2.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
